-- arrayiterators work
git-svn-id: svn://svn.xp-framework.net/xp/trunk@9441 d2cacbed-c0f6-0310-851a-9dad52fc623a

+++ experiments/people/bayer/turpitude/src/native/include/turpitude_jarray.h
@@ -22,6 +22,6 @@ typedef struct turpitude_javaarray_object {
 
 void make_turpitude_jarray();
 void make_turpitude_jarray_instance(jarray array, turpitude_java_type type, zval* dest);
-void turpitude_jarray_get(turpitude_javaarray_object* arr, int idx, zval* return_value);
+void turpitude_jarray_get(turpitude_javaarray_object* arr, unsigned long idx, zval* return_value);
 
 #endif
